Second laptop

+++ araha/mapper/UpdateSubscriptionTypes.go
@@ -1,8 +1,8 @@
 package mapper
 
-import "araha/araha/models"
+import "araha/models"
 
-func UpdateSubscription(subscription models.Subscription) float64 {
+func UpdateSubscription(subscription models.Subscription) int {
 	subscriptionFound := FindSubscriptionTypes(subscription)
 	updatedSubscription := subscriptionFound.Amount + subscription.Amount
 	return updatedSubscription
